Improved the DICOM support: now you can select which way you want the diagnostics statements to be handled. Old version would at new statements now there is a configuration item that allows you to fine tune the result.
Configuration item is called "Mortara Diagnostics" and correct values are: True, False, StateOnly, or HistoryOnly. The value true is default and in this case the system works exactly as it always did the othere values should be self explanatory.

Thanks go to VJP for requesting this changes, and pointing out that the original implementation might not be desirable in all cases.